Rename utilities submodule due to import issues
