+++ pyvista/utilities/__init__.py
@@ -5,4 +5,4 @@ from .fileio import *
 from .geometric_objects import *
 from .parametric_objects import *
 from .sphinx_gallery import Scraper, _get_sg_image_scraper
-from .utilities import *
+from .helpers import *
